+++ rpc/src/v1/impls/personal.rs
@@ -19,13 +19,11 @@ use std::sync::{Arc, Weak};
 use jsonrpc_core::*;
 use v1::traits::Personal;
 use v1::types::TransactionRequest;
-use util::bytes::ToPretty;
+use v1::impls::sign_and_dispatch;
 use util::keys::store::*;
 use util::numbers::*;
-use util::rlp::encode;
 use ethcore::client::BlockChainClient;
-use ethcore::transaction::{Action, SignedTransaction, Transaction as EthTransaction};
-use ethminer::{AccountDetails, MinerService};
+use ethminer::MinerService;
 
 /// Account management (personal) rpc implementation.
 pub struct PersonalClient<A, C, M>
@@ -44,51 +42,6 @@ impl<A, C, M> PersonalClient<A, C, M>
 			client: Arc::downgrade(client),
 			miner: Arc::downgrade(miner),
 		}
-	}
-
-	fn dispatch_transaction(&self, signed_transaction: SignedTransaction) -> Result<Value, Error> {
-		let hash = signed_transaction.hash();
-
-		let import = {
-			let client = take_weak!(self.client);
-			let miner = take_weak!(self.miner);
-
-			miner.import_own_transaction(&*client, signed_transaction, |a: &Address| {
-				AccountDetails {
-					nonce: client.nonce(&a),
-					balance: client.balance(&a),
-				}
-			})
-		};
-
-		match import {
-			Ok(_) => to_value(&hash),
-			Err(e) => {
-				warn!("Error sending transaction: {:?}", e);
-				to_value(&H256::zero())
-			}
-		}
-	}
-
-	fn sign_and_dispatch(&self, request: TransactionRequest, secret: H256) -> Result<Value, Error> {
-		let signed_transaction = {
-			let client = take_weak!(self.client);
-			let miner = take_weak!(self.miner);
-			EthTransaction {
-				nonce: request.nonce
-					.or_else(|| miner
-							 .last_nonce(&request.from)
-							 .map(|nonce| nonce + U256::one()))
-					.unwrap_or_else(|| client.nonce(&request.from)),
-					action: request.to.map_or(Action::Create, Action::Call),
-					gas: request.gas.unwrap_or_else(|| miner.sensible_gas_limit()),
-					gas_price: request.gas_price.unwrap_or_else(|| miner.sensible_gas_price()),
-					value: request.value.unwrap_or_else(U256::zero),
-					data: request.data.map_or_else(Vec::new, |b| b.to_vec()),
-			}.sign(&secret)
-		};
-		trace!(target: "miner", "send_transaction: dispatching tx: {}", encode(&signed_transaction).to_vec().pretty());
-		self.dispatch_transaction(signed_transaction)
 	}
 }
 
@@ -130,7 +83,7 @@ impl<A: 'static, C: 'static, M: 'static> Personal for PersonalClient<A, C, M>
 			.and_then(|(request, password)| {
 				let accounts = take_weak!(self.accounts);
 				match accounts.locked_account_secret(&request.from, &password) {
-					Ok(secret) => self.sign_and_dispatch(request, secret),
+					Ok(secret) => sign_and_dispatch(&self.client, &self.miner, request, secret),
 					Err(_) => to_value(&H256::zero()),
 				}
 		})
